+++ src/main/java/org/ilgcc/app/inputs/Gcc.java
@@ -55,4 +55,6 @@ public class Gcc extends FlowInputs {
     private String ccapStartDate;
 
     private List<String> childcareWeeklySchedule;
+
+    private String current_uuid;
 }
