+++ src/main/resources/messages.properties
@@ -184,8 +184,7 @@ children-ccap-info.race-ethnicity-help-text=Select all that apply. This is optio
 children-ccap-in-care.title=CCAP in care
 children-ccap-in-care.header=Is {0} already in child care with your chosen provider?
 #children-ccap-start-date
-children-ccap-start-date.when-did-title=Previous CCAP Start Date
-children-ccap-start-date.when-will-title=Future CCAP Start Date
+children-ccap-start-date.when-did-title=CCAP Start Date
 children-ccap-start-date.when-did-header=When did {0} start care at your chosen provider?
 children-ccap-start-date.when-will-header=When will {0} start care at your chosen provider?
 #children-childcare-weekly-schedule
