--- src/main/resources/templates/gcc/children-ccap-start-date-previous.html
@@ -1,20 +1,0 @@
-<th:block
-    th:replace="~{fragments/screens/screenWithOneInput ::
-  screenWithOneInput(
-    title=#{children-ccap-start-date.when-did-title},
-    header=#{children-ccap-start-date.when-did-header(${fieldData.get('childFirstName')})},
-    buttonLabel=#{general.button.continue},
-    iconFragment=~{fragments/gcc-icons :: baby},
-    formAction=${formAction},
-    inputContent=~{::inputContent})}">
-  <th:block th:ref="inputContent">
-    <th:block th:replace="~{fragments/inputs/date ::
-    date(inputName='ccapStart',
-    label='',
-    ariaLabel='header',
-    groupName='ccapStartDate')}"/>
-  </th:block>
-</th:block>
-
-
-
