+++ src/main/resources/templates/gcc/children-ccap-start-date.html
@@ -1,0 +1,25 @@
+<div th:with="childHasProvider=${fieldData.get('childInCare').equals('true')},
+              hasProviderHeader=#{children-ccap-start-date.when-did-header(${fieldData.get('childFirstName')})},
+              noCurrentProviderHeader=#{children-ccap-start-date.when-will-header(${fieldData.get('childFirstName')})}">
+  <th:block
+      th:replace="~{fragments/screens/screenWithOneInput ::
+  screenWithOneInput(
+    title=#{children-ccap-start-date.when-did-title},
+    header=${childHasProvider ? hasProviderHeader : noCurrentProviderHeader},
+    buttonLabel=#{general.button.continue},
+    iconFragment=~{fragments/gcc-icons :: baby},
+    formAction=${formAction},
+    inputContent=~{::inputContent})}">
+    <th:block th:ref="inputContent">
+      <input type="hidden" th:name="current_uuid" th:value="${fieldData.get('uuid')}">
+      <th:block th:replace="~{fragments/inputs/date ::
+    date(inputName='ccapStart',
+    label='',
+    ariaLabel='header',
+    groupName='ccapStartDate')}"/>
+    </th:block>
+  </th:block>
+</div>
+
+
+
